+++ lib/browser/browser.rb
@@ -24,6 +24,7 @@ require "browser/weibo"
 require "browser/qq"
 require "browser/alipay"
 require "browser/electron"
+require "browser/facebook"
 
 require "browser/bot"
 require "browser/middleware"
@@ -51,6 +52,7 @@ module Browser
       Edge,
       InternetExplorer,
       Firefox,
+      Facebook,             # must be placed before Chrome and Safari
       Weibo,                # must be placed before Chrome and Safari
       QQ,                   # must be placed before Chrome and Safari
       Alipay,               # must be placed before Chrome and Safari